requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
requests
beautifulsoup4
redis
python-dotenv
colorama
